file and variable naming conventions for type-defs
