+++ src/apollo/server/type-defs/press-release.ts
@@ -1,6 +1,6 @@
 import gql from "graphql-tag";
 
-export const typeDefs = gql`
+export const pressReleaseTypeDefs = gql`
   type PressRelease {
     id: ID!
     title: String!
